data_real/pyproject.toml
[project]
name = "data-real"
version = "0.1.0"
description = "Submodule to execute regression on real system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "hebi-py>=2.12.0",
    "matplotlib>=3.10.8",
    "numpy>=2.3.1",
    "pandas>=2.3.3",
    "py-xl-sindy>=2.2.1",
    "scipy>=1.16.3",
]
